Update README.md
Android Studio, I created a kotlin app for a friend to assist with chosing meals for herself. 
When using the app there are specific time inputs that she has to use and it will display the meal suggestion for her.
The app consists of 6 time inputs and they all have a different suggestion. When entering anything other than the time inputs specified she will be instructed to enter the correct input
The app had two buttons; start and reset. 
After choosing the correct time input you click the start button to display a suggestion, after that you can use the reset button to clear the input and enter a different one.